# openapi.yaml
openapi: 3.0.0
info:
  version: '1'
  title: buildit
  description: REST API spec for Build-it
servers:
- url: "https://build-it-lake.vercel.app/"
paths:
  /api/zoning:
    post:
      summary: Test zoning information
      description: Retrieve zoning information for a given address.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                address:
                  type: string
                  description: The address to test zoning for
      responses:
        '200':
          description: Zoning information for the given address
          content:
            application/json:
              schema:
                type: object
                properties:
                  address:
                    type: string
                  coordinates:
                    type: object
                    properties:
                      lat:
                        type: number
                      lon:
                        type: number
                  zoningCode:
                    type: string
                  zoningInfo:
                    type: object
        '400':
          description: Invalid address
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...